+++ cli/src/web/webAppIcon.ts
@@ -1,12 +1,11 @@
 import {Project} from "../project";
 import * as path from "path";
-import {execute, isDir, makeDirs} from "../utils";
+import {isDir, makeDirs} from "../utils";
 import {rmdirSync} from "fs";
+import {ekc} from "../ekc";
 
 export function webBuildAppIcon(ctx: Project, output: string) {
     const marketAsset = ctx.market_asset ? ctx.market_asset : "assets/res";
-    const binEKC = path.join(ctx.path.EKX_ROOT, "editor/bin/ekc");
-
     if (isDir(output)) {
         rmdirSync(output, {recursive: true});
     }
@@ -21,5 +20,5 @@ export function webBuildAppIcon(ctx: Project, output: string) {
         const scale = size / originalSize;
         cmd.push(scale.toString(), size.toString(), size.toString(), "0", "1", path.join(output, `icon${size}.png`));
     }
-    execute(binEKC, cmd);
+    ekc(ctx, ...cmd);
 }